// src/main/webapp/app/shared/models/contract.model.ts
export interface ContractRaw {
  kalender_id: number;
  produktname: string;
  beschreibung: string | null;
  datum_von: string;
  datum_bis: string;
  kalendar_art: string;
  prodprod_id: number;
  prodspiel_id: number | null;
  prodwerk_id: number | null;
  abko_termin_id: number;
  termin_art: number;
  betrieb_typ: number | null;
  termin_typ: number;
  anba_id: number;
  veranstalter_benutzer: boolean;
  created_benutzer_id: number;
  created_benutzer_Name: string;
  anbieter_betrieb_id: number;
  anbieter_betrieb_betriebsname: string;
  anbieter_betrieb_strasse: string;
  anbieter_betrieb_plz: string;
  anbieter_betrieb_stadt: string;
  veranstalter_betrieb_id: number;
  veranstalter_betrieb_betriebsname: string;
  veranstalter_betrieb_strasse: string;
  veranstalter_betrieb_plz: string;
  veranstalter_betrieb_stadt: string;
  spielss_id: number | null;
  spiels_name: string | null;
  created_am: string;
  veranstaltungskategorie_id: string;
  anbieter_betrieb_email: string;
  veranstalter_betrieb_email: string;
  veranstalter_betrieb_telefon: string;
  anbieter_betrieb_telefon: string;
  vertrag_status: number;
  termin_status: number;
  buch_id: number;
  vertrag_art: number;
  vertrag_anleger_betrieb_id: number;
  vertrag_freigabe_betrieb_id: number;
  buch_vertrag_status: number;
  anbieter_betrieb_typ: number;
  veranstalter_betrieb_typ: number;
  buch_vertrag_ID: number;
  buch_nachtrag_id: number | null;
  storno_status: number;
  spielort_PLZ: string;
  spielort_Stadt: string;
  spielort_Strasse: string;
  spielort_Land: string;
}
export interface Contract {
  vertrag_art: number;
  storno_status: number;
  buch_vertrag_ID: number;
  buch_vertrag_status: number;
  statusLabel: string;
  produktname: string;
  datumsListe: string;
  spielort_Stadt: string;
  vertragspartner: string;
  anbieterId: number;
  prodProdsId: number;
  partnerEmail: string;
  partnerPhone: string;
  adresse: string;
  id: number;
  image: string | null;
  loadingImage: boolean;
  vertrag_anleger_betrieb_id: number;
  vertrag_freigabe_betrieb_id: number;
  termin_typ: number;
  termin_art: number;
  vertrag_status: number;
  benutzerId : string;
  benutzerName : string;
  dateUhrList: Date[];
  abko_termin_id: number;
  buch_nachtrag_id: number | null;

}